pyproject.toml
[project]
name = "cv-tailor"
version = "0.1.0"
description = "An AI-powered tool to tailor your CV (Markdown) to any job description using Google Gemini 3 Flash (Preview)."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=0.1.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
]
